# Conflicts:
#	public/js/dist/all.js
#	public/js/dist/all.js.map
#	public/mix-manifest.json

+++ app/Http/Controllers/ReportTemplatesController.php
@@ -19,7 +19,8 @@ class ReportTemplatesController extends Controller
 
         $report = $request->user()->reportTemplates()->create([
             'name' => $validated['name'],
-            'options' => $request->except(['_token', 'name']),
+            'options' => $request->except(['_token', 'name', 'is_shared']),
+            'is_shared' => $request->has('is_shared'),
         ]);
 
         session()->flash('success', trans('admin/reports/message.create.success'));
@@ -45,6 +46,12 @@ class ReportTemplatesController extends Controller
     {
         $this->authorize('reports.view');
 
+        if ($reportTemplate->created_by != auth()->id()) {
+            return redirect()
+                ->route('report-templates.show', $reportTemplate)
+                ->withError(trans('general.report_not_editable'));
+        }
+
         return view('reports/custom', [
             'customfields' => CustomField::get(),
             'template' => $reportTemplate,
@@ -55,13 +62,29 @@ class ReportTemplatesController extends Controller
     {
         $this->authorize('reports.view');
 
-        // Ignore "options" rules since data does not come in under that key...
-        $validated = $request->validate(Arr::except((new ReportTemplate)->getRules(), 'options'));
+        if ($reportTemplate->created_by != auth()->id()) {
+            return redirect()
+                ->route('report-templates.show', $reportTemplate)
+                ->withError(trans('general.report_not_editable'));
+        }
 
-        $reportTemplate->update([
-            'name' => $validated['name'],
-            'options' => $request->except(['_token', 'name']),
-        ]);
+        $properties = [
+            'name' => $request->input('name'),
+            'options' => Arr::except($request->all(), ['_token', 'id', 'name', 'is_shared']),
+            'is_shared' => $reportTemplate->is_shared,
+        ];
+
+        if ($reportTemplate->created_by == $request->user()->id) {
+            $properties['is_shared'] = $request->boolean('is_shared');
+        }
+
+        $reportTemplate->fill($properties);
+
+        if ($reportTemplate->isInvalid()) {
+            return redirect()->back()->withInput()->withErrors($reportTemplate->getErrors());
+        }
+
+        $reportTemplate->save();
 
         session()->flash('success', trans('admin/reports/message.update.success'));
 
@@ -71,6 +94,12 @@ class ReportTemplatesController extends Controller
     public function destroy(ReportTemplate $reportTemplate): RedirectResponse
     {
         $this->authorize('reports.view');
+
+        if ($reportTemplate->creator()->isNot(auth()->user())) {
+            return redirect()
+                ->route('report-templates.show', $reportTemplate)
+                ->withError(trans('general.generic_model_not_found', ['model' => 'report template']));
+        }
 
         $reportTemplate->delete();
